backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pyjwt==2.8.0
openai==1.3.5
qdrant-client==1.9.2
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
google-generativeai==0.4.1
pytest==7.4.3
lxml==5.1.0